+++ pyroomacoustics/directivities/analytic.py
@@ -59,6 +59,7 @@ a pattern with arbitrary parameter :math:`p`.
 """
 import numpy as np
 
+from .. import random
 from ..doa import spher2cart
 from ..utilities import all_combinations, requires_matplotlib
 from .base import Directivity
@@ -121,6 +122,12 @@ class CardioidFamily(Directivity):
 
         self._gain = gain
         self._pattern_name = f"cardioid family, p={self._p}"
+
+        # this is the object that will allow to sample rays according to the
+        # distribution corresponding to the source energy directivity
+        self.energy_distribution = CardioidEnergyDistribution(
+            loc=self._orientation.unit_vector, p=self._p
+        )
 
     @property
     def is_impulse_response(self):
@@ -199,6 +206,9 @@ class CardioidFamily(Directivity):
                 return np.abs(resp)
             else:
                 return resp
+
+    def sample_rays(self, n_rays, rng=None):
+        return self.energy_distribution.sample(n_rays, rng=rng).T
 
     @requires_matplotlib
     def plot_response(
@@ -443,3 +453,63 @@ def cardioid_energy(p, gain=1.0):
         The gain of the cardioid function
     """
     return gain**2 * (4.0 * np.pi / 3.0) * (4 * p**2 - 2 * p + 1)
+
+
+class CardioidEnergyDistribution(random.distributions.Distribution):
+    """
+    This object draws samples from a cardioid shaped distribution on the sphere
+
+    Parameters
+    ----------
+    loc: array_like or None
+        The unit vector pointing in the main direction of the cardioid.
+        If None, the location defaults to [1, 0, 0].
+    p: float or None
+        Parameter of the cardioid pattern. A value of 0 corresponds to a
+        figure-eight pattern, 0.5 to a cardioid pattern, and 1 to an omni
+        pattern. When None, p is set to 0.5.
+        The parameter must be between 0 and 1
+    """
+
+    def __init__(self, loc=None, p=None):
+        super().__init__(dim=3)
+        if loc is None:
+            loc = np.array([1.0, 0.0, 0.0])
+
+        loc_norm = np.linalg.norm(loc)
+        if abs(loc_norm - 1.0) > 1e-5:
+            raise ValueError(
+                "The location parameter should be a unit "
+                f"norm vector (got norm={loc_norm})."
+            )
+
+        self._loc = loc
+
+        if not (0.0 <= p <= 1.0):
+            raise ValueError(f"The value of p should be between 0.0 and 1.0 (got {p=})")
+
+        self._coeff = p
+
+        self._sampler = random.sampler.RejectionSampler(
+            desired_func=self._unnormalized_pdf,
+            proposal_dist=random.distributions.UnnormalizedUniformSpherical(dim=3),
+            scale=1.0,
+        )
+
+    def _unnormalized_pdf(self, x):
+        x = np.moveaxis(x, -1, 0)
+        response = cardioid_func(
+            x,
+            direction=self._loc,
+            p=self._coeff,
+            gain=1.0,
+            normalize=False,
+            magnitude=True,
+        )
+        return response**2
+
+    def pdf(self, x):
+        return self._unnormalized_pdf(x) / cardioid_energy(p=self._coeff)
+
+    def sample(self, size=None, rng=None):
+        return self._sampler(size=size, rng=rng)