+++ pyroomacoustics/directivities/base.py
@@ -92,3 +92,25 @@ class Directivity(abc.ABC):
             Response at provided angles.
         """
         raise NotImplementedError
+
+    @abc.abstractmethod
+    def sample_rays(self, n_rays, rng=None):
+        """
+        This method samples unit vectors from the sphere according to
+        the distribution of the source
+
+        Parameters
+        ----------
+        n_rays: int
+            The number of rays to sample
+        rng: numpy.random.Generator or None, optional
+            A random number generator object from numpy or None.
+            If None is passed numpy.random.default_rng is used to create
+            a Generator object.
+
+        Returns
+        -------
+        ray_directions: numpy.ndarray, shape (n_dim, n_rays)
+            An array containing the unit vectors in its columns
+        """
+        raise NotImplementedError