+++ pyroomacoustics/libroom_src/libroom.cpp
@@ -65,17 +65,29 @@ PYBIND11_MODULE(libroom, m) {
       .def("get_max_distance", &Room<3>::get_max_distance)
       .def("next_wall_hit", &Room<3>::next_wall_hit)
       .def("scat_ray", &Room<3>::scat_ray)
-      .def("simul_ray", &Room<3>::simul_ray)
+      .def("simul_ray",
+           (void (Room<3>::*)(float phi, float theta,
+                              const Vectorf<3> &source_pos, float energy_0)) &
+               Room<3>::simul_ray)
+      .def("simul_ray",
+           (void (Room<3>::*)(const Vectorf<3> &ray_direction,
+                              const Vectorf<3> &source_pos, float energy_0)) &
+               Room<3>::simul_ray)
       .def("ray_tracing",
-           (void(Room<3>::*)(
+           (void (Room<3>::*)(
                const Eigen::Matrix<float, 2, Eigen::Dynamic> &angles,
-               const Vectorf<3> source_pos)) &
+               const Vectorf<3> &source_pos)) &
                Room<3>::ray_tracing)
-      .def("ray_tracing", (void(Room<3>::*)(size_t nb_phis, size_t nb_thetas,
-                                            const Vectorf<3> source_pos)) &
+      .def("ray_tracing",
+           (void (Room<3>::*)(
+               const Eigen::Matrix<float, 3, Eigen::Dynamic> &vectors,
+               const Vectorf<3> &source_pos)) &
+               Room<3>::ray_tracing)
+      .def("ray_tracing", (void (Room<3>::*)(size_t nb_phis, size_t nb_thetas,
+                                             const Vectorf<3> &source_pos)) &
                               Room<3>::ray_tracing)
       .def("ray_tracing",
-           (void(Room<3>::*)(size_t nb_rays, const Vectorf<3> source_pos)) &
+           (void (Room<3>::*)(size_t nb_rays, const Vectorf<3> &source_pos)) &
                Room<3>::ray_tracing)
       .def("contains", &Room<3>::contains)
       .def_property("is_hybrid_sim", &Room<3>::get_is_hybrid_sim,
@@ -86,6 +98,7 @@ PYBIND11_MODULE(libroom, m) {
       .def_readonly("sources", &Room<3>::sources)
       .def_readonly("orders", &Room<3>::orders)
       .def_readonly("orders_xyz", &Room<3>::orders_xyz)
+      .def_readonly("source_directions", &Room<3>::source_directions)
       .def_readonly("attenuations", &Room<3>::attenuations)
       .def_readonly("gen_walls", &Room<3>::gen_walls)
       .def_readonly("visible_mics", &Room<3>::visible_mics)
@@ -113,17 +126,29 @@ PYBIND11_MODULE(libroom, m) {
       .def("get_max_distance", &Room<2>::get_max_distance)
       .def("next_wall_hit", &Room<2>::next_wall_hit)
       .def("scat_ray", &Room<2>::scat_ray)
-      .def("simul_ray", &Room<2>::simul_ray)
+      .def("simul_ray",
+           (void (Room<2>::*)(float phi, float theta,
+                              const Vectorf<2> &source_pos, float energy_0)) &
+               Room<2>::simul_ray)
+      .def("simul_ray",
+           (void (Room<2>::*)(const Vectorf<2> &ray_direction,
+                              const Vectorf<2> &source_pos, float energy_0)) &
+               Room<2>::simul_ray)
       .def("ray_tracing",
-           (void(Room<2>::*)(
+           (void (Room<2>::*)(
                const Eigen::Matrix<float, 1, Eigen::Dynamic> &angles,
-               const Vectorf<2> source_pos)) &
+               const Vectorf<2> &source_pos)) &
                Room<2>::ray_tracing)
-      .def("ray_tracing", (void(Room<2>::*)(size_t nb_phis, size_t nb_thetas,
-                                            const Vectorf<2> source_pos)) &
+      .def("ray_tracing",
+           (void (Room<2>::*)(
+               const Eigen::Matrix<float, 2, Eigen::Dynamic> &vectors,
+               const Vectorf<2> &source_pos)) &
+               Room<2>::ray_tracing)
+      .def("ray_tracing", (void (Room<2>::*)(size_t nb_phis, size_t nb_thetas,
+                                             const Vectorf<2> &source_pos)) &
                               Room<2>::ray_tracing)
       .def("ray_tracing",
-           (void(Room<2>::*)(size_t n_rays, const Vectorf<2> source_pos)) &
+           (void (Room<2>::*)(size_t n_rays, const Vectorf<2> &source_pos)) &
                Room<2>::ray_tracing)
       .def("contains", &Room<2>::contains)
       .def_property_readonly_static("dim",
@@ -134,6 +159,7 @@ PYBIND11_MODULE(libroom, m) {
       .def_readonly("sources", &Room<2>::sources)
       .def_readonly("orders", &Room<2>::orders)
       .def_readonly("orders_xyz", &Room<2>::orders_xyz)
+      .def_readonly("source_directions", &Room<2>::source_directions)
       .def_readonly("attenuations", &Room<2>::attenuations)
       .def_readonly("gen_walls", &Room<2>::gen_walls)
       .def_readonly("visible_mics", &Room<2>::visible_mics)
@@ -225,16 +251,18 @@ PYBIND11_MODULE(libroom, m) {
 
   // The microphone class
   py::class_<Microphone<3>>(m, "Microphone")
-      .def(py::init<const Vectorf<3> &, int, float, float>())
-      .def_readonly("loc", &Microphone<3>::loc)
-      .def_readonly("hits", &Microphone<3>::hits)
-      .def_readonly("histograms", &Microphone<3>::histograms);
+    .def(py::init<const Vectorf<3> &, int, float, float>())
+    .def("set_directions", &Microphone<3>::set_directions)
+    .def("log_histogram", (void(Microphone<3>::*)(float, const Eigen::ArrayXf &, const Vectorf<3> &))&Microphone<3>::log_histogram)
+    .def_readonly("loc", &Microphone<3>::loc)
+    .def_readonly("histograms", &Microphone<3>::histograms)
+    ;
 
   py::class_<Microphone<2>>(m, "Microphone2D")
-      .def(py::init<const Vectorf<2> &, int, float, float>())
-      .def_readonly("loc", &Microphone<2>::loc)
-      .def_readonly("hits", &Microphone<2>::hits)
-      .def_readonly("histograms", &Microphone<2>::histograms);
+    .def(py::init<const Vectorf<2> &, int, float, float>())
+    .def_readonly("loc", &Microphone<2>::loc)
+    .def_readonly("histograms", &Microphone<2>::histograms)
+    ;
 
   // The 2D histogram class
   py::class_<Histogram2D>(m, "Histogram2D")
@@ -252,8 +280,10 @@ PYBIND11_MODULE(libroom, m) {
       .def_readonly("distance", &Hit::distance);
 
   // getter and setter for geometric epsilon
-  m.def("set_eps", [](const float &eps) { libroom_eps = eps; });
-  m.def("get_eps", []() { return libroom_eps; });
+  m.def("set_eps", [](const float &eps) {
+    libroom_eps = eps; });
+  m.def("get_eps", []() {
+    return libroom_eps; });
 
   // Routines for the geometry packages
   m.def("ccw3p", &ccw3p, "Determines the orientation of three points");
@@ -280,7 +310,7 @@ PYBIND11_MODULE(libroom, m) {
         "Computes the angle between two 2D or 3D vectors");
 
   m.def("dist_line_point", &dist_line_point,
-        "Computes the distance between a point and an infinite line");
+      "Computes the distance between a point and an infinite line");
 
   m.def("rir_builder", &rir_builder, "RIR builder");
   m.def("delay_sum", &delay_sum, "Delay and sum");